cleanup unused code
